requirements.txt
langchain
langchain-community
langchain-openai
langgraph
tavily-python
yfinance
python-dotenv
pydantic
chromadb
sec-edgar-downloader
streamlit
pypdf
tiktoken